Parcial1/requirements.txt
boto3
pytesseract
Pillow